+++ RELEASE_NOTES.md
@@ -4,7 +4,8 @@
 
 - **Table Validators**: Added additional "Script Validators" section for the table validator preferences. The new "Controller.stop" validator checks if a ROM based table with no nvram file, has a table exit routine without a "Controller.stop" call (quite special). Calling this routine there is required for the serialization of the nvram file which contains the tables highscores. Note that this validator requires a new table scan to be triggered. So far I've found "Checkpoint" and the older "X-Files" table affected by this.
 - **VPS Tables Overview**: Added combo box to filter all VPS entries by their table format. The VPX table format is used as default. 
-- **Table Statistics**: Added emulator combo box so that all statistics dashboards can be filtered by the available emulators. 
+- **Table Statistics**: Added emulator combo box so that all statistics dashboards can be filtered by the available emulators.
+- **Table Overview / Playlists**: Added global and local favorites to the list of available playlists in the playlists combo box.
 
 ## Bugfixes