+++ RELEASE_NOTES.md
@@ -12,3 +12,5 @@
 - **Auto-Shutdown**: Fixed resetting of auto-shutdown on key events. This reset got lost when changing to the new controller input API. (And I had such a great start on "Attack From Mars" when the shutdown warning popped up). 
 - **Highscores**: Fixed highscore resolving for **White Water** by adding "RIVER MASTER" to the list of highscore titles.
 - **Highscores**: Fixed ROM resolving and highscore parsing for **American Most Haunted**.
+- **Highscores**: Fixed highscore parsing for **Gemini**, which can now be differed from **Gemini 2000** which is not supported.
+- **iScored**: Fixed missing score submissions, caused by a missing cache reset of the iScored game room by the server.