+++ RELEASE_NOTES.md
@@ -1,4 +1,4 @@
-## Release Notes 3.2.0
+## Release Notes 3.3.0
 
 ## Changes